fix: Movec functions into seperate files
Another attempt at adding a body. Mainly just split up the functions inside set up child data into seperate files as there were way more then 5 functions haha.
